Add Z3 32-bit and 64-bit fp specs
Separate the Z3 FP SAT specs into two, one forcing 32-bit floating
points, the other using the default 64-bit constructs.

+++ z3_fp_sat/set_meta_tests_z3_64.hpp
@@ -23,7 +23,7 @@ namespace zero {
     z3::expr
     get_zero(z3::context& c, z3::expr e)
     {
-        return c.fpa_val((float)0.0);
+        return c.fpa_val(0.0);
     }
 
     z3::expr
@@ -55,7 +55,7 @@ namespace one {
     z3::expr
     get_one(z3::context& c, z3::expr f)
     {
-        return c.fpa_val((float) 1.0);
+        return c.fpa_val(1.0);
     }
 
     z3::expr
